add PriceChangeDetector and PriceVariationPolicy for flight price monitoring

+++ src/main/java/com/maal/searchservice/domain/repository/FlightRepository.java
@@ -1,13 +1,13 @@
 package com.maal.searchservice.domain.repository;
 
 import com.maal.searchservice.domain.modal.FlightPrice;
-import com.maal.searchservice.infra.persistence.entity.WatchRouteEntity;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface FlightRepository {
     void save(FlightPrice flightPrice);
     Optional<FlightPrice> findById(Long flightId);
-    List<FlightPrice> findByPriceBetweenMinAndMax(String origin, String destination, String travelDate, double minPrice, double maxPrice);
+    List<FlightPrice> findByPriceBetweenMinAndMax(String origin, String destination, String travelDate, BigDecimal minPrice, BigDecimal maxPrice);
 }
